:zap: Additional message format checking. Closes #37
Additional checking to incoming network messages has been added and an
'ERR\n' is sent if the message is structured improperly.

The implemented solution isn't the most elegant, but hey, we're kind of in
a hurry. The project must be submitted within 3hrs and the quizzes are
still not yet finished. Also we gotta do the "tutorials" on how to run the
files. Why am I even typing this? This shouldn't be a part of the commit
message. Also, apparently it will rain towmorrow.